requirements.txt
Django>=5.0
pandas
PyMuPDF
PyPDF2
tabula-py
openjdk-11-jre-headless